add send message, get and delete chat endpoints

+++ internal/db/models/message.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type MessageModel struct {
-	ID        int64     `gorm:"primaryKey"`
-	ChatID    int64     `gorm:"not null;index"`
+	ID        int       `gorm:"primaryKey"`
+	ChatID    int       `gorm:"not null;index"`
 	Text      string    `gorm:"type:text;not null;check:char_length(text) <= 5000"`
 	CreatedAt time.Time `gorm:"not null;default:now()"`
 }
